+++ packages/rainbowkit/src/wallets/connectorsForWallets.ts
@@ -1,5 +1,5 @@
 import { Connector } from 'wagmi';
-import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
+import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
 import { isHexString } from '../utils/colors';
 import { isMobile } from '../utils/isMobile';
 import { omitUndefinedValues } from '../utils/omitUndefinedValues';
@@ -102,7 +102,7 @@ export const connectorsForWallets = (walletList: WalletList) => {
         ) {
           const { chains, options } = connector;
 
-          walletConnectModalConnector = new WalletConnectConnector({
+          walletConnectModalConnector = new WalletConnectLegacyConnector({
             chains,
             options: {
               ...options,