+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -1,10 +1,12 @@
 import { argentWallet } from './argentWallet/argentWallet';
+import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
 import { metaMaskWallet } from './metaMaskWallet/metaMaskWallet';
+import { mewWallet } from './mewWallet/mewWallet';
 import { omniWallet } from './omniWallet/omniWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
 import { trustWallet } from './trustWallet/trustWallet';
@@ -12,12 +14,14 @@ import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
 
 export {
   argentWallet,
+  bitskiWallet,
   braveWallet,
   coinbaseWallet,
   imTokenWallet,
   injectedWallet,
   ledgerWallet,
   metaMaskWallet,
+  mewWallet,
   omniWallet,
   rainbowWallet,
   trustWallet,