+++ examples/with-next-siwe-iron-session/CHANGELOG.md
@@ -1,5 +1,28 @@
 # with-next-siwe-iron-session
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [3399df5]
+- Updated dependencies [8f01a12]
+  - @rainbow-me/rainbowkit@0.12.1
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [9838acf]
+  - @rainbow-me/rainbowkit@0.12.0
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [924ae82]
+- Updated dependencies [5e233ea]
+  - @rainbow-me/rainbowkit@0.11.1
+
 ## 0.0.17
 
 ### Patch Changes